Make StackedAreaChart a reusable component
